Added service interfaces and refactored existing services to be implementations of those services.
