+++ hotel-booking-app-service/src/main/java/com/application/hotelbooking/services/HotelService.java
@@ -1,43 +1,8 @@
 package com.application.hotelbooking.services;
 
-import com.application.hotelbooking.domain.RoomModel;
 import com.application.hotelbooking.dto.HotelCreationServiceDTO;
-import com.application.hotelbooking.exceptions.InvalidHotelException;
-import com.application.hotelbooking.services.repositoryservices.HotelRepositoryService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
 
-import java.util.Comparator;
-import java.util.Optional;
-
-@Service
-public class HotelService {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(HotelService.class);
-
-    @Autowired
-    private HotelRepositoryService hotelRepositoryService;
-
-    public void createHotel(HotelCreationServiceDTO hotelCreationServiceDTO ){
-        if (isHotelNameFree(hotelCreationServiceDTO)){
-            hotelRepositoryService.save(hotelCreationServiceDTO);
-        } else {
-            throw new InvalidHotelException("That hotelName is already taken");
-        }
-    }
-
-    private boolean isHotelNameFree(HotelCreationServiceDTO hotelCreationServiceDTO) {
-        return hotelRepositoryService.findHotelByHotelName(hotelCreationServiceDTO.getHotelName()).isEmpty();
-    }
-
-    public int getLatestRoomNumberOfHotel(Long hotelId) {
-        return hotelRepositoryService.getHotelById(hotelId).get()
-                .getRooms()
-                .stream()
-                .mapToInt(RoomModel::getRoomNumber)
-                .max()
-                .orElse(0);
-    }
+public interface HotelService {
+    void createHotel(HotelCreationServiceDTO hotelCreationServiceDTO);
+    int getLatestRoomNumberOfHotel(Long hotelId);
 }
