+++ hotel-booking-app-web/src/main/java/com/application/hotelbooking/controllers/AddRoomsController.java
@@ -1,7 +1,6 @@
 package com.application.hotelbooking.controllers;
 
 import com.application.hotelbooking.dto.RoomCreationDTO;
-import com.application.hotelbooking.exceptions.InvalidRoomException;
 import com.application.hotelbooking.services.RoomService;
 import com.application.hotelbooking.services.repositoryservices.HotelRepositoryService;
 import com.application.hotelbooking.transformers.HotelViewTransformer;
